Fix and improve quick reply and quick reply emote parser
1) Migrates jQuery.get calls to use the Fetch API instead. Fetch
is compatible with all modern browsers SALR Redux supports. This
enables the emote parser to work successfully in Firefox.
Additional logic is used to feature detect the presence of
content.fetch on Firefox 58+, which sends the fetch request as if
it were from the content page instead of from the extension. This
ensures that credentials are set despite any third-party cookie
denial settings the user may have chosen.

2) Moves jQuery's parsing of fetched reply/edit pages to a new
document instead of implicitly being parsed in the same document.
This slightly enhances extension security and prevents the
boogeyman from appearing in Firefox <=57 with third-party cookies
disabled.

3) Fixes a typo in EmoteParser that led to cached emotes never
being updated. (this.expireTimeOut vs this.expireTimeout)

4) Updates emote parser to fetch over https to prevent man-in-the
middle attacks.

5) Fixes two bugs in quick edit so it works on single post pages.
It was throwing an exception since it couldn't find a bookmark
star on the page, so it now just presets it to false and fetches
the actual status from the reply page. The same technology can be
added to quick quote/reply if it's preferred to fetch the current
state instead of whatever the state was when the page was loaded.

6) Adds a check so we don't store missing/malformed emote data.

7) Last and least, cleans up some missing semicolons.