Firefox compatibility updates (#64)
* Update chrome messaging API calls to use chrome.runtime instead

(Thanks, Serfer!)

This improves compatibility with Firefox. Since 2012, the
chrome.extension calls were just aliases for chrome.runtime,
so this shouldn't introduce any chrome compatibility issues.

* Update manifest to use options_ui instead of options_page

(Thanks, Serfer!)

This improves compatibility with Firefox. The open_in_tab
flag preserves previous extension behavior w.r.t. options_page.
options_ui is the preferred to set this since Chrome 40.

Chrome documentation claims open_in_tab will eventually be
removed, but in this case preserving behavior is more important.

* Update hotkey-manager to use event.which instead of event.keycode

(Thanks, Serfer!)

This improves compatibility with Firefox. Unfortunately, both
of the above are deprecated - if the functionality gets removed,
this should be redone to support the replacement event.key.

* Improve the page navigator

- Adds a new setting to make the last post button reload the
    current thread to check for new posts instead of scrolling
    to an unread post. The setting defaults to false, preserving
    the previous behavior for Chrome. For the Firefox version,
    this should default to true as that is what the Firefox users
    are familiar with/expecting.
- Cleans up the page navigator CSS. With a preset width, there
    was some obnoxious empty space. The pre-set width and CSS
    that set a negative margin on #nav-body have been removed.
    Padding has been added to the right side, and the left-hand
    side's padding was slightly improved. Also, the pesky tiny
    underlines between page navigator buttons that had links
    have been removed.
- Hides the last post button in the page navigator for pages
    where the button did nothing.

* Fix what appears to be an unintentionally-global variable

The variable is not referenced outside of the function it's in,
so it was most likely just an accident.

* Update handlers to use tabs.query instead of tabs.getSelected

This improves compatibility with Firefox. tabs.getSelected is
deprecated.

* Fix quick quote so it actually works on single post pages

Improves findThreadID() by adding another means of finding thread
IDs for showthread pages. This fetches the thread ID from "Show
posts by this user" links that appear in each post, which are the
only known way to find the thread ID from a single post page.

With this change, submitting a quick quote reply on a single post
page will no longer produce a 'Specified thread was not found in
live forums' error.

Also fixes an undeclared var.

* Fix up message handlers for image uploading

Moves message listener for appending images to quick reply.
Sends the info for the image to append back to the tab that
actually uploaded the image instead of whichever tab the user
happened to be browsing at the time.

Also cleans up a few variable declarations.

* Fix and improve quick reply and quick reply emote parser

1) Migrates jQuery.get calls to use the Fetch API instead. Fetch
is compatible with all modern browsers SALR Redux supports. This
enables the emote parser to work successfully in Firefox.
Additional logic is used to feature detect the presence of
content.fetch on Firefox 58+, which sends the fetch request as if
it were from the content page instead of from the extension. This
ensures that credentials are set despite any third-party cookie
denial settings the user may have chosen.

2) Moves jQuery's parsing of fetched reply/edit pages to a new
document instead of implicitly being parsed in the same document.
This slightly enhances extension security and prevents the
boogeyman from appearing in Firefox <=57 with third-party cookies
disabled.

3) Fixes a typo in EmoteParser that led to cached emotes never
being updated. (this.expireTimeOut vs this.expireTimeout)

4) Updates emote parser to fetch over https to prevent man-in-the
middle attacks.

5) Fixes two bugs in quick edit so it works on single post pages.
It was throwing an exception since it couldn't find a bookmark
star on the page, so it now just presets it to false and fetches
the actual status from the reply page. The same technology can be
added to quick quote/reply if it's preferred to fetch the current
state instead of whatever the state was when the page was loaded.

6) Adds a check so we don't store missing/malformed emote data.

7) Last and least, cleans up some missing semicolons.