Fix Quick Reply-related hotkeys
1) Adds missing `event` parameter to `formatText`. Firefox didn't
have a reference there.

2) Changes `event.srcElement` to `event.target`; `srcElement` was
an alias to `target` on Chrome but not Firefox. With this and the
above change, quick reply formatting hotkeys work in Firefox.

3) Migrates from the deprecated `event.keyCode` to the current
`event.key`. Note that `event.key` is case-sensitive and I've
picked the lower-case version to use here.

4) Moves some declarations and definitions that are only used
inside a conditional so they're also inside that conditional.

5) Fixes hotkey-manager so it correctly handles the output of
`findCurrentPage()`. This makes the quick reply-related kb-nav
hotkeys work on both Firefox and Chrome.

6) Adds conditionals so we don't try to open quick reply if it's
disabled, and also so we don't open unread threads if those are
disabled.

Also includes some trivial cleanup in hotkey-manager.