New: Avatar toggling and user post background colors
Initial work on creating a master function to loop through posts
once in showthread. So far, it includes SAARS and avatar-hiding.

Adds hide/show avatar toggle button. Setting for this is enabled
by default. The hidden avatar list is stored as a stringified
array in localStorage.

Adds UI to the user note option to set a background color for
a user's posts. User note backgrounds are, for now, set as
important priority to override mod/admin backgrounds. This is
only temporary and will be resolved when user notes and mod/admin
background-setting are combined into the master showthread
function.

Adds protection against setting empty color CSS if a color isn't
specified for a user note. It might be worth changing the default
note color in the future from red to unset.