Fix load new posts with last post button breaking on page 1
Also fixes it to work in single post view.

The original logic required the existence of an ampersand in the
URL. This didn't always happen, so trying to jump to any new
unread posts would fail. Instead of regex, we now build the
new URL with what we have in the location object.

Cleans up a couple of unguarded for-in loops in util.