Roll plat grenade icon hiding into handleShowThread's purview
Simplifies the logic to only remove the class if there aren't
any other special classes attached to the user. The previous
behavior still showed the diamond/pill icons when the platinum
class was removed; the new behavior for those is to still show
them but also keep the 'platinum' class. This way, we won't need
to maintain a list of special cases or re-add any CSS.

This also fixes the bug where, with that option enabled, my
username would be completely obscured by my icon.