Improve BBCode/Emote insertion
New behavior:
BBCode: If text is selected, it is surrounded by tags.
        If no text is selected, tags are inserted at the cursor.
Emotes: If text is selected, the emote is added afterwards.
        If no text is selected, the emote is added at the cursor.

Focus is restored to the message box after clicking these
elements so users can see where things will go or what
existing text will be affected by adding a tag.

Replaces jQuery event delegation with equivalent JS.

We can consider adding this same technology to the hotkeys if
we don't mind changing what is selected afterwards - the hotkeys
currently keep the selection, while the new `insertText`
function moves the text cursor to the end of the added text.