Improve image uploader
Adds option to thumbnail uploaded images. If the box is checked,
the result is placed in `timg` tags instead of `img` tags. The
checkbox defaults to unchecked, but will remember its value as
long as the image upload panel remains open.

Hooks up the invisible input element to the uploader. Adds a type
filter to its HTML so only imgur-accepted images may be selected.
Adds a UI hint that the drop zone text can be clicked.

Makes text in the image upload panel unselectable.

Also does some minor cleanup.